Move tests out of subdirectory
